requirements.txt
blinker==1.4
Click==7.0
Flask==1.1.1
Flask-DebugToolbar==0.10.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
six==1.12.0
SQLAlchemy==1.3.7
Werkzeug==0.15.5
